move socket tests to socket
